Prevent crashes in padding & pooling layers.
A crash occurred when the input image was smaller than the output image.
This was fixed by augmenting the util.pad_nearest_pow2 with a threshold that must be exceeded.
ImageRescaler will now pad any arbitrary input image to be a power of two (that is at least the size of the output!) before applying pooling.
The number of candidates is now explicitly passed to the ballot recognizer instead of being a configurable hyperparameter.