Swap classification for bag of labels.
Assigning each vote a class has been swapped for assigning each vote a bag of selected labels.
This allows our NN to handle the case where to voter voted on nothing, or they voted for multiple candidates.
However, I don't know how to judge accuracy under this model.
We will need to discuss how to tell if a ballot was recorded accurately.