Make CNN configurable via hyperparameters.
The convolutional/pooling/fully connected layers of ImageRecognizer may now be configured through the config object.
Moved extremely long and repeated lines of code to utils.

Increased the minimum size of each contest in the y direction.
Contests were allowed to be smaller than the first convolutional filter, which caused the NN to perform *very* poorly.