Create 2-stage ballot marking process: generate & apply.
Ballot marking now consists of two steps: generation and application.

Generation takes in contest image and a bounding rectangle for an area to be marked.
It returns an array that is meant to be overlayed on the bounding rectangle on the contest image.

Application takes an image overlay and combines it with the contest image.
The most straightforward application is to assign the overlay directly to the underlying pixels.
A more complex application may stretch, skew, or adjust the center of the overlay.

With these two features, we should be able to synthesize realistic ballots.