Rasterize Oregon ballots, reduce ballot sizes.
Each marked ballot maintained a pointer to the ballot template object, and marked contest maintained a pointer to the contest definition.
These pointers vastly inflated the size of the object (48M->1K),=.
Upon analysis of the code, these pointer were unecessary, and thus removed.

If you need information about the contests on a marked ballot, you will have to pass along the template object.
This reduces the pickled size of our marked ballots immensely.

test_raster was modified so that it correctly cropped option bounding rectangles as well as contest rectangles.
Assertions about the pixel coordinates should enforce the output of the module.