Create BallotFactory to control ballot lifetimes.
Managing ballot lifetimes (and assigning ballot indecies) was impossible in prior versions.
This made it difficult to create a voting evaluation system with multiple ballots.

Now the DataSet should be converted to either accept a ballot factory and populate it (generative datasets).
Alternatively, the dataset would read from the disk and yield an instantiated. ballot factory.