Process multiple ballots in contest recognition.
The neural network code has been updated to process multiple ballot definitions simultaneously.

Initial results are lackluster in terms of accuracy, but the code doesn't crash.

Future improvements are as follows:

1) More CNNs.
Create multiple CNN's inside the image recognizer.
In training mode, the input to the NN gets run through *each* CNN, and a tensor containing *all* results is returned.
All results are forwarded through all output labelers.
The size of the optimizer step is scaled by some function (either 1/(best-worst), something involving atan), or only k of n are updated.
Each CNN should learn some characteristics of a particular contest.
In evaluation mode, all of the data is fed through the NN, and index of the "best" CNN for each (ballot, contest) pair is recorded.
The evaluation results are based off of the "best" CNN.
With unlabeled testing data, only the result from the best CNN is returned.

2) Changing data batching.
Re-do the data loaders so it emits single contests rather than entire ballots.
The dataset would need to ensure that output labels are always the same length FOR ALL CONTESTS.
Images must be of different length, so a custom_collate_fn would need to be provided to the DataLoader that uses stack() rather than cat().
The (ragged) image tensor would be un-batched within rescaler, and per-image rescaling would be applied.
The CNN would proceed as normal.
The output would change drastically.
There would only be one output layer, with as many outputs as the maximum number of options for one contest.
An additional input to this stage would be the (ballot, contest) index.
This FC network would need to be huge for it to learn to process each ballot individually.

3) One network per ballot
In this approach, we would abandon this multi-ballot system.
Each ballot would get one old-style NN, and no sharing would occur between CNN's or output layers.
This is guaranteed to work (well), but the size fo the model (and training time) would scale linearly with the number of ballot definitions.