Add configuration lines for generator network.
Added flags that allow you to implement your own generator in the GAN.
Do note that the generator is picky and will not like most configurations.
This is because the input AND output sizes of the GAN are fixed.
So, you will have to repeatedly experiment with GAN configurations until the sizes work out.
Documentation forthcoming.