Create Auto-Encoder from GAN generator.
After conversation with Dr. Garman, we decided to construct an auto-encoder that uses our gan as the decoder stage.
This will be useful in diagnosing difficulties in our NN.
Since our image quality is low, this will help us determine if the generator is bad or if the discriminator is bad.
Or, it could determine that the tensor-to-image conversion doesn't work.